+++ release/files_common
@@ -385,6 +385,7 @@ selfdrive/camerad/imgproc/utils.cc
 selfdrive/camerad/imgproc/utils.h
 
 selfdrive/manager/__init__.py
+selfdrive/manager/helpers.py
 selfdrive/manager/process.py
 selfdrive/manager/manager.py
